refactor: introduce class InnerBodyCreationDelegates
