+++ src/M31.FluentApi.Generator/CodeGeneration/CodeBoardActors/InnerBodyGeneration/LineForMethodGenerator.cs
@@ -21,7 +21,7 @@ internal class LineForMethodGenerator : LineGeneratorBase<MethodSymbolInfo>
     protected override void GenerateLineWithoutReflection(MethodSymbolInfo symbolInfo)
     {
         CallMethodCode callMethodCode = new CallMethodCode(BuildCallMethodCode, CodeBoard.NewLineString);
-        CodeBoard.AssignCallMethodCode(symbolInfo, callMethodCode);
+        CodeBoard.InnerBodyCreationDelegates.AssignCallMethodCode(symbolInfo, callMethodCode);
 
         List<string> BuildCallMethodCode(string instancePrefix, IReadOnlyCollection<Parameter> outerMethodParameters)
         {
@@ -43,7 +43,7 @@ internal class LineForMethodGenerator : LineGeneratorBase<MethodSymbolInfo>
     protected override void GenerateLineWithReflection(MethodSymbolInfo symbolInfo, string infoFieldName)
     {
         CallMethodCode callMethodCode = new CallMethodCode(BuildCallMethodCode, CodeBoard.NewLineString);
-        CodeBoard.AssignCallMethodCode(symbolInfo, callMethodCode);
+        CodeBoard.InnerBodyCreationDelegates.AssignCallMethodCode(symbolInfo, callMethodCode);
 
         List<string> BuildCallMethodCode(string instancePrefix, IReadOnlyCollection<Parameter> outerMethodParameters)
         {
