+++ src/M31.FluentApi.Generator/CodeGeneration/CodeBoardElements/CodeBoard.cs
@@ -1,7 +1,6 @@
 using M31.FluentApi.Generator.CodeBuilding;
 using M31.FluentApi.Generator.CodeGeneration.CodeBoardActors.Commons;
 using M31.FluentApi.Generator.CodeGeneration.CodeBoardActors.MethodCreation.Forks;
-using M31.FluentApi.Generator.Commons;
 using M31.FluentApi.Generator.SourceAnalyzers;
 using M31.FluentApi.Generator.SourceGenerators;
 using M31.FluentApi.Generator.SourceGenerators.AttributeElements;
@@ -14,8 +13,6 @@ namespace M31.FluentApi.Generator.CodeGeneration.CodeBoardElements;
 internal class CodeBoard
 {
     private readonly List<Diagnostic> diagnostics;
-    private readonly Dictionary<string, SetMemberCode> memberToSetMemberCode;
-    private readonly Dictionary<MethodIdentity, CallMethodCode> methodToCallMethodCode;
 
     private CodeBoard(
         IReadOnlyCollection<FluentApiInfo> fluentApiInfos,
@@ -33,8 +30,7 @@ internal class CodeBoard
         BuilderClass = builderClass;
         Constructor = null;
         StaticConstructor = null;
-        memberToSetMemberCode = new Dictionary<string, SetMemberCode>();
-        methodToCallMethodCode = new Dictionary<MethodIdentity, CallMethodCode>();
+        InnerBodyCreationDelegates = new InnerBodyCreationDelegates();
         BuilderMethodToAttributeData = new Dictionary<BuilderMethod, AttributeDataExtended>();
         Forks = new List<Fork>();
         BuilderClassFields = new BuilderClassFields();
@@ -50,8 +46,7 @@ internal class CodeBoard
     internal Class BuilderClass { get; }
     internal Method? Constructor { get; set; }
     internal Method? StaticConstructor { get; set; }
-    internal IReadOnlyDictionary<string, SetMemberCode> MemberToSetMemberCode => memberToSetMemberCode;
-    internal IReadOnlyDictionary<MethodIdentity, CallMethodCode> MethodToCallMethodCode => methodToCallMethodCode;
+    internal InnerBodyCreationDelegates InnerBodyCreationDelegates { get; }
     internal Dictionary<BuilderMethod, AttributeDataExtended> BuilderMethodToAttributeData { get; }
     internal IReadOnlyList<Fork> Forks { get; set; }
     internal BuilderClassFields BuilderClassFields { get; }
@@ -113,43 +108,6 @@ internal class CodeBoard
         codeFile.AddHeaderLine("// This code was generated by the library M31.FluentAPI.");
         codeFile.AddHeaderLine("// Changes to this file may cause incorrect behavior and will be lost if the code is " +
                                "regenerated.");
-    }
-
-    internal void AssignSetMemberCode(MemberSymbolInfo memberSymbolInfo, SetMemberCode setMemberCode)
-    {
-        string memberName = memberSymbolInfo.Name;
-
-        if (memberToSetMemberCode.ContainsKey(memberName))
-        {
-            throw new GenerationException(
-                $"{nameof(SetMemberCode)} for member {memberName} has already been assigned.");
-        }
-
-        memberToSetMemberCode[memberName] = setMemberCode;
-    }
-
-    internal void AssignCallMethodCode(MethodSymbolInfo methodSymbolInfo, CallMethodCode callMethodCode)
-    {
-        MethodIdentity methodIdentity = CreateMethodIdentity(methodSymbolInfo);
-
-        if (methodToCallMethodCode.ContainsKey(methodIdentity))
-        {
-            throw new GenerationException(
-                $"{nameof(CallMethodCode)} for method {methodIdentity.MethodName} has already been assigned.");
-        }
-
-        methodToCallMethodCode[methodIdentity] = callMethodCode;
-    }
-
-    internal CallMethodCode GetCallMethodCode(MethodSymbolInfo methodSymbolInfo)
-    {
-        return methodToCallMethodCode[CreateMethodIdentity(methodSymbolInfo)];
-    }
-
-    private static MethodIdentity CreateMethodIdentity(MethodSymbolInfo methodSymbolInfo)
-    {
-        return MethodIdentity.Create(methodSymbolInfo.Name,
-            methodSymbolInfo.ParameterInfos.Select(i => i.TypeForCodeGeneration));
     }
 
     internal void ReportDiagnostic(Diagnostic diagnostic)
